Implement Renderable interface in RandomSearch class
The RandomSearch class now implements the Renderable interface, allowing it to provide grid data for rendering. Also, static keyword removed for methods and variables within the class. New Renderable interface was added to Phase3.JFX3D package. This modification brings about better encapsulation and follows the design principle of programming to an interface, not an implementation.