Add centralized Settings.java class, update module name
Introduced a Settings.java class to centralize configuration for window, camera, and cube settings in the JFX3D project. Adjusted access modifier of the RandomSearch.java's method to private and updated the module name in module-info.java from MyModule to Application. Effort is taken to make configuration management easier and more modular.