Fixed old rendering bug
Amir's greedy search was static. Which means that it is initialised once when the program starts and attempting to create new instances of it will just return the old pre-existing instance. That is why the output grid data stayed the same despite the change in databases.

PLEASE TRY TO NOT USE STATIC CLASSES IF AT ALL POSSIBLE.