Reorder scene setup methods in JFX3D
The scene setup methods in the JFX3D class have been reordered to fix an existing rendering bug. The menu scene setup is now located above the render stage setup. Previously, the old static instance of the search was used which led to grid data inconsistencies when changing databases. This change helps to avoid such problems.