Refactor code to use AlgoRequest and AlgoResponse
The data method in Renderable class is updated to accept an AlgoRequest object and return an AlgoResponse object. Removed getData method that takes Updatable as a second parameter in the Renderable interface. AlgoRequest and AlgoResponse classes are added to facilitate this change. Also, JFX3D.java is updated to handle these changes.